+++ CHANGELOG.md
@@ -9,6 +9,10 @@ New features:
 - Add `PresenceRunner::on_retry` for accepting a closure to execute during socket creation / handshake retries. The returned `usize` value (through the closure parameter) indicates the amount of total retries done at the time of the closure's execution.
 - Add `PresenceRunner::set_max_retries` for enabling developers to set a custom retry count for socket creation / handshakes after which the runner instance should give up on connecting to Discord. By default this is set to `0` (indefinite) internally.
 
+Changes:
+
+- `PresenceClient::client_id` now returns an `&str` instead of a cloned `String`.
+
 Internal changes / improvements:
 
 - Removed the `json!()` macro call from `DiscordSock::do_handshake` and replaced it with `PresenceHandshake<'_>`.